requirements.txt
asn1crypto==0.24.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3.1
diff-match-patch==20121119
Django==2.0.6
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-filter==2.0.0
django-formtools==2.1
django-guardian==1.4.9
django-import-export==1.0.1
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-redis==4.9.0
django-simple-captcha==0.5.7
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
DjangoUeditor==1.8.143
drf-extensions==0.3.1
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.11.3
idna==2.7
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
mysqlclient==1.3.12
odfpy==1.3.6
openpyxl==2.5.7
Pillow==5.2.0
pycparser==2.18
pycryptodome==3.6.5
PyJWT==1.6.4
PyMySQL==0.8.1
pyOpenSSL==18.0.0
pypiwin32==223
pytz==2018.5
pywin32==223
PyYAML==3.13
redis==2.10.6
requests==2.19.1
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.23
xlrd==1.1.0
xlwt==1.3.0
yunpian-python-sdk==1.0.0
